requirements.txt
Django>=4.2
requests
gunicorn>=21.2.0
python-dotenv>=1.0.0
